# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Base Eleitoral 360 API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /liderancas:
    get:
      operationId: listLiderancas
      tags: [liderancas]
      summary: List all leaderships
      responses:
        "200":
          description: List of leaderships
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Lideranca"
    post:
      operationId: createLideranca
      tags: [liderancas]
      summary: Create a leadership
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LiderancaInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lideranca"
  /liderancas/{id}:
    get:
      operationId: getLideranca
      tags: [liderancas]
      summary: Get a leadership by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Leadership
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lideranca"
    patch:
      operationId: updateLideranca
      tags: [liderancas]
      summary: Update a leadership
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LiderancaUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lideranca"
    delete:
      operationId: deleteLideranca
      tags: [liderancas]
      summary: Delete a leadership
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── Apoiadores ──────────────────────────────────────────────────────────────
  /apoiadores:
    get:
      operationId: listApoiadores
      tags: [apoiadores]
      summary: List all supporters
      responses:
        "200":
          description: List of supporters
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Apoiador"
    post:
      operationId: createApoiador
      tags: [apoiadores]
      summary: Create a supporter
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApoiadorInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Apoiador"
  /apoiadores/{id}:
    get:
      operationId: getApoiador
      tags: [apoiadores]
      summary: Get a supporter by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Supporter
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Apoiador"
    patch:
      operationId: updateApoiador
      tags: [apoiadores]
      summary: Update a supporter
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApoiadorUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Apoiador"
    delete:
      operationId: deleteApoiador
      tags: [apoiadores]
      summary: Delete a supporter
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── Prospecção ──────────────────────────────────────────────────────────────
...
  /prospeccao/{id}/status:
    patch:
      operationId: updateProspeccaoStatus
      tags: [prospeccao]
      summary: Update prospect pipeline stage
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProspeccaoStatusUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Prospecto"

  # ─── Zonas Eleitorais ────────────────────────────────────────────────────────
...
  /agenda:
    get:
      operationId: listAgenda
      tags: [agenda]
      summary: List field events
      responses:
        "200":
          description: Events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EventoAgenda"
    post:
      operationId: createEvento
      tags: [agenda]
      summary: Create a field event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventoAgendaInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventoAgenda"
  /agenda/{id}:
    patch:
      operationId: updateEvento
      tags: [agenda]
      summary: Update a field event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventoAgendaUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventoAgenda"
    delete:
      operationId: deleteEvento
      tags: [agenda]
      summary: Delete a field event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── Demandas ────────────────────────────────────────────────────────────────
  /demandas:
    get:
      operationId: listDemandas
      tags: [demandas]
      summary: List population demands
      responses:
        "200":
          description: Demands
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Demanda"
    post:
      operationId: createDemanda
      tags: [demandas]
      summary: Create a demand
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DemandaInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Demanda"
  /demandas/{id}:
    patch:
      operationId: updateDemanda
      tags: [demandas]
      summary: Update a demand
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DemandaUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Demanda"
    delete:
      operationId: deleteDemanda
      tags: [demandas]
      summary: Delete a demand
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── Mapas ───────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Lideranca:
      type: object
      required: [id, nome, bairro, cidade, estado, tipoLideranca, status, grauConfianca]
      properties:
        id:
          type: integer
        nome:
          type: string
        apelido:
          type: ["string", "null"]
        telefone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        tipoLideranca:
          type: string
        status:
          type: string
        cep:
          type: ["string", "null"]
        rua:
          type: ["string", "null"]
        numero:
          type: ["string", "null"]
        complemento:
          type: ["string", "null"]
        bairro:
          type: string
        cidade:
          type: string
        estado:
          type: string
        regiaoAtuacao:
          type: ["string", "null"]
        liderancaSuperior:
          type: ["string", "null"]
        responsavelInterno:
          type: ["string", "null"]
        apoiadoresCadastrados:
          type: integer
        apoiadoresEstimadosDiretos:
          type: integer
        apoiadoresEstimadosIndiretos:
          type: integer
        votosDeclarados:
          type: integer
        votosValidados:
          type: integer
        grauConfianca:
          type: string
        fonteEstimativa:
          type: ["string", "null"]
        comprovacao:
          type: ["string", "null"]
        ultimaAtualizacao:
          type: ["string", "null"]
        proximaAcao:
          type: ["string", "null"]
        observacoes:
          type: ["string", "null"]
    LiderancaInput:
      type: object
      required: [nome, bairro, cidade, estado, tipoLideranca, status, grauConfianca]
      properties:
        nome:
          type: string
        apelido:
          type: string
        telefone:
          type: string
        email:
          type: string
        tipoLideranca:
          type: string
        status:
          type: string
        cep:
          type: string
        rua:
          type: string
        numero:
          type: string
        complemento:
          type: string
        bairro:
          type: string
        cidade:
          type: string
        estado:
          type: string
        regiaoAtuacao:
          type: string
        liderancaSuperior:
          type: string
        responsavelInterno:
          type: string
        apoiadoresCadastrados:
          type: integer
        apoiadoresEstimadosDiretos:
          type: integer
        apoiadoresEstimadosIndiretos:
          type: integer
        votosDeclarados:
          type: integer
        votosValidados:
          type: integer
        grauConfianca:
          type: string
        fonteEstimativa:
          type: string
        comprovacao:
          type: string
        proximaAcao:
          type: string
        observacoes:
          type: string
    LiderancaUpdate:
      type: object
      properties:
        nome:
          type: string
        apelido:
          type: string
        telefone:
          type: string
        email:
          type: string
        tipoLideranca:
          type: string
        status:
          type: string
        cep:
          type: string
        rua:
          type: string
        numero:
          type: string
        complemento:
          type: string
        bairro:
          type: string
        cidade:
          type: string
        estado:
          type: string
        regiaoAtuacao:
          type: string
        liderancaSuperior:
          type: string
        responsavelInterno:
          type: string
        apoiadoresCadastrados:
          type: integer
        apoiadoresEstimadosDiretos:
          type: integer
        apoiadoresEstimadosIndiretos:
          type: integer
        votosDeclarados:
          type: integer
        votosValidados:
          type: integer
        grauConfianca:
          type: string
        fonteEstimativa:
          type: string
        comprovacao:
          type: string
        proximaAcao:
          type: string
        observacoes:
          type: string

    # ─── Apoiadores ────────────────────────────────────────────────────────────
    Apoiador:
      type: object
      required: [id, nome, bairro, cidade, estado, tipoPessoa, statusPolitico, nivelConfianca]
      properties:
        id:
          type: integer
        nome:
          type: string
        telefone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        cep:
          type: ["string", "null"]
        rua:
          type: ["string", "null"]
        numero:
          type: ["string", "null"]
        complemento:
          type: ["string", "null"]
        bairro:
          type: string
        cidade:
          type: string
        estado:
          type: string
        liderancaVinculada:
          type: ["string", "null"]
        tipoPessoa:
          type: string
        statusPolitico:
          type: string
        nivelConfianca:
          type: string
        nivelPrecisaoGeografica:
          type: ["string", "null"]
        observacoes:
          type: ["string", "null"]
        dataUltimoContato:
          type: ["string", "null"]
        proximaAcao:
          type: ["string", "null"]
        responsavelContato:
          type: ["string", "null"]
    ApoiadorInput:
      type: object
      required: [nome, bairro, cidade, estado, tipoPessoa, statusPolitico, nivelConfianca]
      properties:
        nome:
          type: string
        telefone:
          type: string
        email:
          type: string
        cep:
          type: string
        rua:
          type: string
        numero:
          type: string
        complemento:
          type: string
        bairro:
          type: string
        cidade:
          type: string
        estado:
          type: string
        liderancaVinculada:
          type: string
        tipoPessoa:
          type: string
        statusPolitico:
          type: string
        nivelConfianca:
          type: string
        nivelPrecisaoGeografica:
          type: string
        observacoes:
          type: string
        dataUltimoContato:
          type: string
        proximaAcao:
          type: string
        responsavelContato:
          type: string
    ApoiadorUpdate:
      type: object
      properties:
        nome:
          type: string
        telefone:
          type: string
        email:
          type: string
        cep:
          type: string
        rua:
          type: string
        numero:
          type: string
        complemento:
          type: string
        bairro:
          type: string
        cidade:
          type: string
        estado:
          type: string
        liderancaVinculada:
          type: string
        tipoPessoa:
          type: string
        statusPolitico:
          type: string
        nivelConfianca:
          type: string
        nivelPrecisaoGeografica:
          type: string
        observacoes:
          type: string
        dataUltimoContato:
          type: string
        proximaAcao:
          type: string
        responsavelContato:
          type: string

    # ─── Prospecção ────────────────────────────────────────────────────────────
    Prospecto:
      type: object
      required: [id, nome, etapa, bairro]
      properties:
        id:
          type: integer
        nome:
          type: string
        telefone:
          type: ["string", "null"]
        bairro:
          type: string
        cidade:
          type: ["string", "null"]
        etapa:
          type: string
        responsavel:
          type: ["string", "null"]
        dataContato:
          type: ["string", "null"]
        observacoes:
          type: ["string", "null"]
...
    ProspeccaoStatusUpdate:
      type: object
      required: [etapa]
      properties:
        etapa:
          type: string

    # ─── Zonas Eleitorais ──────────────────────────────────────────────────────
...
    EventoAgenda:
      type: object
      required: [id, titulo, tipo, data, bairro, status]
      properties:
        id:
          type: integer
        titulo:
          type: string
        tipo:
          type: string
        data:
          type: string
        horario:
          type: ["string", "null"]
        bairro:
          type: string
        cidade:
          type: ["string", "null"]
        responsavel:
          type: ["string", "null"]
        status:
          type: string
        observacoes:
          type: ["string", "null"]
    EventoAgendaInput:
      type: object
      required: [titulo, tipo, data, bairro, status]
      properties:
        titulo:
          type: string
        tipo:
          type: string
        data:
          type: string
        horario:
          type: string
        bairro:
          type: string
        cidade:
          type: string
        responsavel:
          type: string
        status:
          type: string
        observacoes:
          type: string
    EventoAgendaUpdate:
      type: object
      properties:
        titulo:
          type: string
        tipo:
          type: string
        data:
          type: string
        horario:
          type: string
        bairro:
          type: string
        cidade:
          type: string
        responsavel:
          type: string
        status:
          type: string
        observacoes:
          type: string

    # ─── Demandas ──────────────────────────────────────────────────────────────
    Demanda:
      type: object
      required: [id, bairro, tipoDemanda, prioridade, status]
      properties:
        id:
          type: integer
        pessoaVinculada:
          type: ["string", "null"]
        bairro:
          type: string
        tipoDemanda:
          type: string
        prioridade:
          type: string
        status:
          type: string
        responsavel:
          type: ["string", "null"]
        observacoes:
          type: ["string", "null"]
        dataRetorno:
          type: ["string", "null"]
    DemandaInput:
      type: object
      required: [bairro, tipoDemanda, prioridade, status]
      properties:
        pessoaVinculada:
          type: string
        bairro:
          type: string
        tipoDemanda:
          type: string
        prioridade:
          type: string
        status:
          type: string
        responsavel:
          type: string
        observacoes:
          type: string
        dataRetorno:
          type: string
    DemandaUpdate:
      type: object
      properties:
        pessoaVinculada:
          type: string
        bairro:
          type: string
        tipoDemanda:
          type: string
        prioridade:
          type: string
        status:
          type: string
        responsavel:
          type: string
        observacoes:
          type: string
        dataRetorno:
          type: string

    # ─── Mapas ─────────────────────────────────────────────────────────────────
...