# lib/api-zod/src/generated/types/demandaInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Base Eleitoral 360 API
 * OpenAPI spec version: 0.1.0
 */

export interface DemandaInput {
  pessoaVinculada?: string;
  bairro: string;
  tipoDemanda: string;
  prioridade: string;
  status: string;
  responsavel?: string;
  observacoes?: string;
  dataRetorno?: string;
}