# lib/api-zod/src/generated/types/eventoAgenda.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Base Eleitoral 360 API
 * OpenAPI spec version: 0.1.0
 */

export interface EventoAgenda {
  id: number;
  titulo: string;
  tipo: string;
  data: string;
  /** @nullable */
  horario?: string | null;
  bairro: string;
  /** @nullable */
  cidade?: string | null;
  /** @nullable */
  responsavel?: string | null;
  status: string;
  /** @nullable */
  observacoes?: string | null;
}