# lib/api-zod/src/generated/types/eventoAgendaInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Base Eleitoral 360 API
 * OpenAPI spec version: 0.1.0
 */

export interface EventoAgendaInput {
  titulo: string;
  tipo: string;
  data: string;
  horario?: string;
  bairro: string;
  cidade?: string;
  responsavel?: string;
  status: string;
  observacoes?: string;
}