# lib/api-zod/src/generated/types/prospecto.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Base Eleitoral 360 API
 * OpenAPI spec version: 0.1.0
 */

export interface Prospecto {
  id: number;
  nome: string;
  /** @nullable */
  telefone?: string | null;
  bairro: string;
  /** @nullable */
  cidade?: string | null;
  etapa: string;
  /** @nullable */
  responsavel?: string | null;
  /** @nullable */
  dataContato?: string | null;
  /** @nullable */
  observacoes?: string | null;
}